(MAINT) - Remove powershell dependency upper bound
Prior to this commit, the mount_iso module had an explicit dependency on
versions 4 and below of the puppetlabs-powershell module.
This caused a dependency conflict when using the puppetlabs-sqlserver module
as it depends on both of these modules, and installed a higher version of
puppetlabs-powershell then was allowed by the mount_iso module.